check if the user has administrative permission on the organism, if not, cannot delete the organism

+++ tools/apollo/delete_organism.py
@@ -9,6 +9,13 @@ from webapollo import AssertUser, GuessOrg, OrgOrGuess, WAAuth, WebApolloInstanc
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
+def checkPermission(gx_user, org_cn):
+    org_permissions = wa.users.getOrganismPermissionsForUser(gx_user)
+    for organism in org_permissions:
+        if org_cn == organism['organism'] and "ADMINISTRATE" in organism['permissions']:
+            return True
+    return False
+
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Sample script to completely delete an organism')
@@ -25,12 +32,16 @@ if __name__ == '__main__':
     # Get organism
     org_cn = GuessOrg(args, wa)
     if org_cn:
-        # TODO: Check user perms on org.
+        # Check user perms on org.
         org_cn = org_cn[0]
         org = wa.organisms.findOrganismByCn(org_cn)
-        returnData = wa.organisms.deleteOrganism(org['id'])
-        print("Delete organism %s" % org_cn)
-        print("returnData = " + str(returnData) + "\n")
+        if checkPermission(gx_user, org_cn):
+            returnData = wa.organisms.deleteOrganism(org['id'])
+            print("Delete organism %s" % org_cn)
+            print("returnData = " + str(returnData) + "\n")
+        else:
+            logger.error("Not authorized! You do not have permissions to delete this organism. You need to request administrative permission for this organism from the owner.")
+            exit(1)
     else:
         logger.error("Organism %s doesn't exist" % org_cn)
